tools/llm/pyproject.toml
[project]
name = "llm-tools"
version = "0.1.0"
description = "LLM client wrappers for GOTCHA framework"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
